requirements.txt
fastapi
uvicorn
docker
openai
python-dotenv
pydantic
pydantic-settings
